Checkpoint:
First of all, a pretty-sure-it's-all-good point for the simplifying the macros.
There's probably some further simplifications of the macros, and definitely
more removal of old cruft, but as it this can be merged back into the trunk.

Second, there's much of the incorporation of BBN's nst-xml-dump here.  Not 100%
complete, but much of the heavy lifting.  By the end of the weekend (but
unfortunately probably not today), this NST can be shipped up to BBN.