Fixed the use of make-check-result.  This was called inside apply, but was a defstruct constructor.  Defstruct constructors are macros and so not eligible for use with APPLY.
In order to make a minimum-complexity fix, I simply added a shim
function, make-check-result that wraps around a renamed constructor
(%make-check-result).  This meant that I didn't need to fix all of the
call sites.

I found one call site that tried to call make-check-result with the
:erring keyword argument (which seemed to be effectively "error
count").  I can't figure out how this /ever/ worked, so I killed it.