Use trivial-backtrace.
For Allegro, this is a little bit "one step backward, not quite with
the two steps forward yet".  I've ripped out the logic that helps
Allegro narrow down the stack to just what's relevant to the test
failing.  So now, ACL will show all the NST set-up that it used to
suppress, plus now trivial-backtrace fun.