Initial coding for individually cached fixtures.
Along with the code change for the feature is a re-write of the def-fixture macro.  Rather than several collect-loops through the fixtures' forms, there's now one loop with several collect-into's and a finally block.

The new cached-fixtures feature has not been tested, but all previous unit tests pass.

The new feature is not yet documented.